Move DP IP address structure into a separate file
