+++ include/dp_firewall.h
@@ -7,10 +7,13 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
+
 #include <sys/queue.h>
 #include <rte_common.h>
+#include "dp_ipaddr.h"
 #include "dp_mbuf_dyn.h"
-#include "dp_util.h"
+
+#define DP_FIREWALL_ID_MAX_LEN	64
 
 #define DP_FWALL_MATCH_ANY_PORT			0xFFFFFFFF
 #define DP_FWALL_MATCH_ANY_ICMP_TYPE	0xFFFFFFFF
