+++ include/dp_ipaddr.h
@@ -1,0 +1,45 @@
+// SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and IronCore contributors
+// SPDX-License-Identifier: Apache-2.0
+
+#ifndef __INCLUDE_DP_IPADDR_H__
+#define __INCLUDE_DP_IPADDR_H__
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+#include <stdint.h>
+#include <stdbool.h>
+
+#define DP_IPV6_ADDR_SIZE 16
+
+struct dp_ip_address {
+	bool			is_v6;
+	union {
+		uint8_t		ipv6[DP_IPV6_ADDR_SIZE];
+		uint32_t	ipv4;
+	};
+};
+
+// inspired by DPDK's RTE_ETHER_ADDR_PRT_FMT and RTE_ETHER_ADDR_BYTES
+// network byte-order!
+#define DP_IPV4_PRINT_FMT       "%u.%u.%u.%u"
+#define DP_IPV4_PRINT_BYTES(ip) (ip) & 0xFF, \
+								((ip) >> 8) & 0xFF, \
+								((ip) >> 16) & 0xFF, \
+								((ip) >> 24) & 0xFF
+
+#define DP_IPV6_PRINT_FMT       "%x:%x:%x:%x:%x:%x:%x:%x"
+#define DP_IPV6_PRINT_BYTES(ip) rte_cpu_to_be_16(((uint16_t *)(ip))[0]), \
+								rte_cpu_to_be_16(((uint16_t *)(ip))[1]), \
+								rte_cpu_to_be_16(((uint16_t *)(ip))[2]), \
+								rte_cpu_to_be_16(((uint16_t *)(ip))[3]), \
+								rte_cpu_to_be_16(((uint16_t *)(ip))[4]), \
+								rte_cpu_to_be_16(((uint16_t *)(ip))[5]), \
+								rte_cpu_to_be_16(((uint16_t *)(ip))[6]), \
+								rte_cpu_to_be_16(((uint16_t *)(ip))[7])
+
+#ifdef __cplusplus
+}
+#endif
+#endif
